chore(xlsx): extract chart dataTable + dataLabels hosts into chart/
Move every parseDataTable / buildDataTable / resolveDataTable* helper
from chart-reader.ts and chart-writer.ts into chart/dataTable.ts.
Move every parseDataLabels / buildDataLabels / resolveDataLabels*
helper plus VALID_DLBL_POSITIONS into chart/dataLabels.ts.

The clone-side resolveDataTable / resolveChartDataLabels /
resolveSeriesDataLabels overrides keep their 3-arg
(source, override) signatures and stay in chart-clone.ts; they
delegate to the shared writer-side normalizers / shape helpers.

Pure relocation — JSDoc preserved verbatim, no behavior change. All
7362 tests still pass; typecheck clean.
