+++ src/xlsx/chart-reader.ts
@@ -99,6 +99,8 @@ import {
   parseLegendStrikethrough,
   parseLegendUnderline,
 } from "./chart/legend";
+import { parseDataTable } from "./chart/dataTable";
+import { parseDataLabels } from "./chart/dataLabels";
 
 /** All chart-type element local names recognized by Excel. */
 const CHART_KIND_TAGS: ReadonlyMap<string, ChartKind> = new Map([
@@ -3036,568 +3038,6 @@ function parseMarker(ser: XmlElement): ChartMarker | undefined {
 
 // ── Data Labels ───────────────────────────────────────────────────
 
-const VALID_DLBL_POSITIONS: ReadonlySet<ChartDataLabelPosition> = new Set([
-  "t",
-  "b",
-  "l",
-  "r",
-  "ctr",
-  "inEnd",
-  "inBase",
-  "outEnd",
-  "bestFit",
-]);
-
-/**
- * Read a `<c:dLbls>` block. Returns `undefined` when the block is
- * empty or only contains a `<c:delete val="1">` (which suppresses
- * labels rather than describing them). All toggles default to `false`
- * when the matching `<c:show*>` element is absent.
- */
-function parseDataLabels(el: XmlElement): ChartDataLabelsInfo | undefined {
-  // <c:delete val="1"> at the root of <c:dLbls> means "suppress for
-  // this scope". We don't surface a dataLabels record for that case —
-  // it's the absence of labels, not a configuration.
-  const deleteEl = findChild(el, "delete");
-  if (deleteEl && readBoolAttr(deleteEl) === true) return undefined;
-
-  const out: ChartDataLabelsInfo = {};
-
-  const pos = findChild(el, "dLblPos");
-  if (pos) {
-    const val = pos.attrs.val;
-    if (typeof val === "string" && VALID_DLBL_POSITIONS.has(val as ChartDataLabelPosition)) {
-      out.position = val as ChartDataLabelPosition;
-    }
-  }
-
-  // `<c:showLegendKey val=".."/>` mirrors Excel's "Format Data Labels
-  // -> Legend Key" checkbox. The OOXML default is `false`, so absence
-  // and an explicit `val="0"` collapse to `undefined` — only an
-  // explicit `val="1"` (or `"true"`) surfaces `true`. Same shape as the
-  // other `show*` toggles so the parsed record can be fed straight back
-  // into {@link cloneChart}.
-  const showLeg = findChild(el, "showLegendKey");
-  if (showLeg && readBoolAttr(showLeg) === true) out.showLegendKey = true;
-
-  const showVal = findChild(el, "showVal");
-  if (showVal && readBoolAttr(showVal) === true) out.showValue = true;
-
-  const showCat = findChild(el, "showCatName");
-  if (showCat && readBoolAttr(showCat) === true) out.showCategoryName = true;
-
-  const showSer = findChild(el, "showSerName");
-  if (showSer && readBoolAttr(showSer) === true) out.showSeriesName = true;
-
-  const showPct = findChild(el, "showPercent");
-  if (showPct && readBoolAttr(showPct) === true) out.showPercent = true;
-
-  const sep = findChild(el, "separator");
-  if (sep) {
-    const text = elementText(sep);
-    if (text.length > 0) out.separator = text;
-  }
-
-  // `<c:numFmt formatCode=".." sourceLinked=".."/>` mirrors Excel's
-  // "Format Data Labels -> Number" panel — pinning a custom number
-  // format on the rendered label values. Same shape as the axis-side
-  // `<c:numFmt>` so the parsed value can be fed straight back into
-  // {@link cloneChart}. The element sits early in the CT_DLbls
-  // sequence (after the optional `<c:dLbl>` instances), so the lookup
-  // is scoped to direct `<c:dLbls>` children — a `<c:numFmt>` nested
-  // inside a per-point `<c:dLbl>` does not leak into the block-level
-  // record.
-  const numFmt = parseDataLabelsNumberFormat(el);
-  if (numFmt) out.numberFormat = numFmt;
-
-  // `<c:showLeaderLines val=".."/>` mirrors Excel's "Format Data
-  // Labels -> Show Leader Lines" checkbox. The OOXML default is
-  // `true` (Excel paints leader lines on every label that gets pushed
-  // outside its slice), so absence and `val="1"` collapse to
-  // `undefined` — only an explicit `val="0"` (or `"false"`) surfaces
-  // `false`. Mirrors how the writer-side scope guard treats the
-  // element: only meaningful on pie / doughnut, but the parser is
-  // permissive (the OOXML schema scopes the element to `EG_DLbls` for
-  // `CT_PieChart` / `CT_DoughnutChart`, but a templated chart whose
-  // type element ends up coerced should still surface the source's
-  // intent so the cloned model stays accurate).
-  // Only the literal OOXML falsy spellings (`"0"` / `"false"`) flip the
-  // toggle — unknown / missing val tokens collapse to `undefined` rather
-  // than surface a `false` the writer would round-trip into a non-default
-  // `<c:showLeaderLines val="0"/>` Excel never authored.
-  const showLeader = findChild(el, "showLeaderLines");
-  if (showLeader) {
-    const v = showLeader.attrs.val;
-    if (typeof v === "string" && (v === "0" || v.toLowerCase() === "false")) {
-      out.showLeaderLines = false;
-    }
-  }
-
-  // `<c:txPr><a:p><a:pPr><a:defRPr sz="N"/></a:pPr></a:p></c:txPr>`
-  // — data-label font size pinned via Excel's "Format Data Labels ->
-  // Font -> Size" knob. The OOXML attribute is in 100ths of a point on
-  // `CT_TextCharacterProperties`' `sz` slot; the reader divides by
-  // 100 at parse time so the surfaced value matches what the user
-  // sees in the UI. Out-of-range / malformed tokens collapse to
-  // `undefined` so absence and a malformed source value round-trip
-  // identically through `cloneChart`.
-  const fontSize = parseDataLabelsFontSize(el);
-  if (fontSize !== undefined) out.fontSize = fontSize;
-
-  // `<c:txPr><a:p><a:pPr><a:defRPr><a:solidFill><a:srgbClr val=".."/>
-  // </a:solidFill></a:defRPr></a:pPr></a:p></c:txPr>` — data-label
-  // font color pinned via Excel's "Format Data Labels -> Font -> Font
-  // color" picker. Theme references (`<a:schemeClr>`) and malformed
-  // `val` tokens collapse to `undefined` since only the literal RGB
-  // triple round-trips losslessly through `writeChart`.
-  const fontColor = parseDataLabelsFontColor(el);
-  if (fontColor !== undefined) out.fontColor = fontColor;
-
-  // `<c:txPr><a:p><a:pPr><a:defRPr b=".."/></a:pPr></a:p></c:txPr>` —
-  // data-label bold flag pinned via Excel's "Format Data Labels ->
-  // Font -> Bold" toggle. Only an explicit `b="1"` surfaces `true`;
-  // the OOXML default `b="0"` collapses to `undefined` so absence
-  // and `b="0"` round-trip identically through `cloneChart`.
-  const bold = parseDataLabelsBold(el);
-  if (bold !== undefined) out.bold = bold;
-
-  // `<c:txPr><a:p><a:pPr><a:defRPr i=".."/></a:pPr></a:p></c:txPr>` —
-  // data-label italic flag pinned via Excel's "Format Data Labels ->
-  // Font -> Italic" toggle. Only an explicit `i="1"` surfaces `true`;
-  // the OOXML default `i="0"` collapses to `undefined` so absence
-  // and `i="0"` round-trip identically through `cloneChart`.
-  const italic = parseDataLabelsItalic(el);
-  if (italic !== undefined) out.italic = italic;
-
-  // `<c:txPr><a:p><a:pPr><a:defRPr u=".."/></a:pPr></a:p></c:txPr>` —
-  // data-label underline flag pinned via Excel's "Format Data Labels
-  // -> Font -> Underline" toggle. Only `u="sng"` (Excel's UI variant
-  // — single underline) surfaces `true`; the OOXML default `"none"`
-  // (and every other ST_TextUnderlineType variant) collapse to
-  // `undefined` so absence and `u="none"` round-trip identically
-  // through `cloneChart`.
-  const underline = parseDataLabelsUnderline(el);
-  if (underline !== undefined) out.underline = underline;
-
-  // `<c:txPr><a:p><a:pPr><a:defRPr strike=".."/></a:pPr></a:p>
-  // </c:txPr>` — data-label strikethrough flag pinned via Excel's
-  // "Format Data Labels -> Font -> Strikethrough" toggle. Only
-  // `strike="sngStrike"` (Excel's UI variant — single line) surfaces
-  // `true`; the OOXML default `"noStrike"` and the non-UI variant
-  // `"dblStrike"` (and any malformed token) collapse to `undefined`
-  // so absence and `"noStrike"` round-trip identically through
-  // `cloneChart`.
-  const strikethrough = parseDataLabelsStrikethrough(el);
-  if (strikethrough !== undefined) out.strikethrough = strikethrough;
-
-  // `<c:txPr><a:p><a:pPr><a:defRPr><a:latin typeface=".."/></a:defRPr>
-  // </a:pPr></a:p></c:txPr>` — data-label font family pinned via
-  // Excel's "Format Data Labels -> Font -> Font" picker. Empty /
-  // whitespace-only `typeface` attributes and missing `<a:latin>`
-  // elements both collapse to `undefined` so absence and the empty
-  // form round-trip identically through the writer.
-  const fontFamily = parseDataLabelsFontFamily(el);
-  if (fontFamily !== undefined) out.fontFamily = fontFamily;
-
-  // `<c:spPr><a:solidFill><a:srgbClr val=".."/></a:solidFill></c:spPr>`
-  // — data-labels background fill pinned via Excel's "Format Data Labels
-  // -> Fill -> Solid fill -> Color" picker. Theme references
-  // (`<a:schemeClr>`), non-solid fills (`<a:noFill>` / `<a:gradFill>` /
-  // `<a:pattFill>` / `<a:blipFill>`), and malformed `val` tokens
-  // collapse to `undefined` since only the literal RGB triple
-  // round-trips losslessly through `writeChart`. Distinct from
-  // `fontColor` (which lives on `<c:txPr><a:p><a:pPr><a:defRPr>
-  // <a:solidFill>`); the two knobs target different children of
-  // `<c:dLbls>` so a caller can pin both without conflict.
-  const fillColor = parseDataLabelsFillColor(el);
-  if (fillColor !== undefined) out.fillColor = fillColor;
-
-  // `<c:spPr><a:ln><a:solidFill><a:srgbClr val=".."/></a:solidFill>
-  // </a:ln></c:spPr>` — data-labels border (line) color pinned via
-  // Excel's "Format Data Labels -> Border -> Solid line -> Color"
-  // picker. Theme references (`<a:schemeClr>`), non-solid line fills
-  // (`<a:noFill>` / `<a:gradFill>` / `<a:pattFill>`), and malformed
-  // `val` tokens collapse to `undefined` since only the literal RGB
-  // triple round-trips losslessly through `writeChart`. Distinct from
-  // `fillColor` (which lives on `<c:spPr><a:solidFill>`); the two
-  // knobs target different children of the same `<c:spPr>` so a
-  // caller can pin both without conflict.
-  const borderColor = parseDataLabelsBorderColor(el);
-  if (borderColor !== undefined) out.borderColor = borderColor;
-
-  // `<c:dLbls><c:spPr><a:ln w="EMU">` carries Excel's "Format Data
-  // Labels -> Border -> Width" pin. Delegates to the shared
-  // {@link parseBorderWidthFromSpPr} so the EMU encoding and snap /
-  // clamp grammar match every other chart-frame border-width slot.
-  const borderWidth = parseBorderWidthFromSpPr(el);
-  if (borderWidth !== undefined) out.borderWidth = borderWidth;
-
-  // `<c:dLbls><c:spPr><a:ln><a:prstDash val=".."/>` carries Excel's
-  // "Format Data Labels -> Border -> Dash type" pin. Delegates to the
-  // shared {@link parseBorderDashFromSpPr} helper.
-  const borderDash = parseBorderDashFromSpPr(el);
-  if (borderDash !== undefined) out.borderDash = borderDash;
-
-  // Empty record is meaningless to a consumer — collapse to undefined.
-  if (
-    out.position === undefined &&
-    !out.showValue &&
-    !out.showCategoryName &&
-    !out.showSeriesName &&
-    !out.showPercent &&
-    !out.showLegendKey &&
-    out.separator === undefined &&
-    out.numberFormat === undefined &&
-    out.showLeaderLines === undefined &&
-    out.fontSize === undefined &&
-    out.fontColor === undefined &&
-    out.bold === undefined &&
-    out.italic === undefined &&
-    out.underline === undefined &&
-    out.strikethrough === undefined &&
-    out.fontFamily === undefined &&
-    out.fillColor === undefined &&
-    out.borderColor === undefined &&
-    out.borderWidth === undefined &&
-    out.borderDash === undefined
-  ) {
-    return undefined;
-  }
-  return out;
-}
-
-/**
- * Pull `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr sz="N"/></a:pPr></a:p>
- * </c:txPr></c:dLbls>` off a data-labels block. Returns the font size
- * in points (`1..400`), or `undefined` when the element is absent /
- * the chain is malformed at any link / the surfaced value is out of
- * the supported range.
- *
- * The OOXML attribute is in 100ths of a point on
- * `CT_TextCharacterProperties`' `sz` slot (ECMA-376 Part 1,
- * §21.1.2.3.7); the reader divides by 100 at parse time so the
- * surfaced value matches what the user sees in Excel's UI. Mirrors
- * the chart-title / axis-title / axis tick-label / legend font-size
- * readers exactly so a parsed value slots straight back into the
- * writer's emit path.
- */
-function parseDataLabelsFontSize(dLbls: XmlElement): number | undefined {
-  const txPr = findChild(dLbls, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.sz;
-  if (typeof raw !== "string") return undefined;
-  const trimmed = raw.trim();
-  if (trimmed.length === 0) return undefined;
-  const parsed = Number.parseInt(trimmed, 10);
-  if (!Number.isFinite(parsed)) return undefined;
-  // Convert from 100ths of a point to points, rounding to the nearest
-  // 0.5pt to match the granularity Excel's UI exposes. Mirrors the
-  // chart-title / axis-title / tick-label / legend sibling parsers
-  // exactly so a parsed value flows through every typography slot
-  // without bookkeeping the units.
-  const halfSteps = Math.round((parsed / TITLE_FONT_SZ_PER_POINT) * 2);
-  const points = halfSteps / 2;
-  if (points < TITLE_FONT_SIZE_MIN_PT || points > TITLE_FONT_SIZE_MAX_PT) return undefined;
-  return points;
-}
-
-/**
- * Pull `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr><a:solidFill>
- * <a:srgbClr val="RRGGBB"/></a:solidFill></a:defRPr></a:pPr></a:p>
- * </c:txPr></c:dLbls>` off a data-labels block. Returns the
- * 6-character uppercase hex string when the parser walks the full
- * chain and lands on an `<a:srgbClr val="RRGGBB"/>`. Theme references
- * (`<a:schemeClr>`), `<a:hslClr>`, `<a:sysClr>`, and `<a:prstClr>`
- * all collapse to `undefined` — only the literal RGB triple
- * round-trips losslessly through {@link writeChart}. Malformed `val`
- * tokens (wrong length, non-hex characters) likewise drop to
- * `undefined` rather than fabricate a value the writer would
- * round-trip into a malformed `<a:srgbClr>`.
- *
- * Returns `undefined` whenever the data-labels block omits `<c:txPr>`
- * entirely or the canonical chain is malformed at any link. Mirrors
- * the chart-title / axis-title / axis tick-label / legend color
- * readers exactly so a parsed value slots straight back into the
- * writer's emit path.
- */
-function parseDataLabelsFontColor(dLbls: XmlElement): string | undefined {
-  const txPr = findChild(dLbls, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const solidFill = findChild(defRPr, "solidFill");
-  if (!solidFill) return undefined;
-  const srgbClr = findChild(solidFill, "srgbClr");
-  if (!srgbClr) return undefined;
-  return normalizeRgbHex(srgbClr.attrs.val);
-}
-
-/**
- * Pull `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr b=".."/></a:pPr></a:p>
- * </c:txPr></c:dLbls>` off a data-labels block. Returns the bold
- * flag.
- *
- * The OOXML `b` attribute is the `xsd:boolean` bold flag on
- * `CT_TextCharacterProperties`. Only an explicit `b="1"` (or
- * `"true"`) surfaces `true`; the OOXML default `0` (and absence /
- * malformed tokens) collapses to `undefined` so absence and the
- * default round-trip identically through `cloneChart`. Mirrors the
- * chart-title / axis-title / axis tick-label / legend bold readers
- * exactly so a parsed value slots straight back into the writer's
- * emit path.
- */
-function parseDataLabelsBold(dLbls: XmlElement): boolean | undefined {
-  const txPr = findChild(dLbls, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.b;
-  if (raw === "1" || raw === "true") return true;
-  return undefined;
-}
-
-/**
- * Pull `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr i=".."/></a:pPr></a:p>
- * </c:txPr></c:dLbls>` off a data-labels block. Returns the italic
- * flag.
- *
- * The OOXML `i` attribute is the `xsd:boolean` italic flag on
- * `CT_TextCharacterProperties`. Only an explicit `i="1"` (or
- * `"true"`) surfaces `true`; the OOXML default `0` (and absence /
- * malformed tokens) collapses to `undefined` so absence and the
- * default round-trip identically through `cloneChart`. Mirrors the
- * chart-title / axis-title / axis tick-label / legend italic readers
- * exactly so a parsed value slots straight back into the writer's
- * emit path.
- */
-function parseDataLabelsItalic(dLbls: XmlElement): boolean | undefined {
-  const txPr = findChild(dLbls, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.i;
-  if (raw === "1" || raw === "true") return true;
-  return undefined;
-}
-
-/**
- * Pull `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr u=".."/></a:pPr></a:p>
- * </c:txPr></c:dLbls>` off a data-labels block. Returns the underline
- * flag.
- *
- * The OOXML `u` attribute is the `ST_TextUnderlineType` enumeration
- * on `CT_TextCharacterProperties`. Only `u="sng"` (Excel's UI variant
- * — single underline) surfaces `true`; the OOXML default `"none"`
- * (and every other variant the schema allows — `"dbl"`, `"heavy"`,
- * `"dotted"`, `"dotDash"`, `"wavy"`, etc.) collapse to `undefined` so
- * absence and `u="none"` round-trip identically through `cloneChart`.
- * Reporting any non-`"sng"` underline as `true` would silently
- * downgrade the choice to a single line on round-trip; the writer
- * emits only `u="sng"` / `u="none"`, matching the boolean shape the
- * UI exposes. Mirrors the chart-title / axis-title / axis tick-label
- * / legend underline readers exactly so a parsed value slots straight
- * back into the writer's emit path.
- */
-function parseDataLabelsUnderline(dLbls: XmlElement): boolean | undefined {
-  const txPr = findChild(dLbls, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.u;
-  if (raw === "sng") return true;
-  return undefined;
-}
-
-/**
- * Pull `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr strike=".."/></a:pPr>
- * </a:p></c:txPr></c:dLbls>` off a data-labels block. Returns the
- * strikethrough flag.
- *
- * The OOXML `strike` attribute is the `ST_TextStrikeType`
- * enumeration on `CT_TextCharacterProperties`. Only
- * `strike="sngStrike"` (Excel's UI variant — single line) surfaces
- * `true`; the OOXML default `"noStrike"` and the non-UI variant
- * `"dblStrike"` (and any malformed token) collapse to `undefined` so
- * absence and `"noStrike"` round-trip identically through
- * `cloneChart`. Reporting `"dblStrike"` as `true` would silently
- * downgrade the choice to a single line on round-trip; the writer
- * emits only `"sngStrike"`, matching the boolean shape the UI
- * exposes. Mirrors the chart-title / axis-title / axis tick-label /
- * legend strikethrough readers exactly so a parsed value slots
- * straight back into the writer's emit path.
- */
-function parseDataLabelsStrikethrough(dLbls: XmlElement): boolean | undefined {
-  const txPr = findChild(dLbls, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.strike;
-  if (raw === "sngStrike") return true;
-  return undefined;
-}
-
-/**
- * Pull `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr><a:latin
- * typeface=".."/></a:defRPr></a:pPr></a:p></c:txPr></c:dLbls>` off a
- * data-labels block. Returns the typeface string the labels were
- * authored with.
- *
- * The OOXML `<a:latin>` element carries the typeface name on
- * `CT_TextFont` (ECMA-376 Part 1, §21.1.2.3.7). The reader trims
- * surrounding whitespace and reports the trimmed typeface; empty /
- * whitespace-only `typeface` attributes and missing `<a:latin>`
- * elements both collapse to `undefined` so absence and the empty
- * form round-trip identically through the writer. Non-string
- * `typeface` tokens (defensive — the XML parser only ever surfaces
- * strings) likewise drop to `undefined`.
- *
- * Returns `undefined` whenever the data-labels block omits `<c:txPr>`
- * entirely or the canonical `<a:p><a:pPr><a:defRPr><a:latin>` chain
- * is malformed at any link. Mirrors the chart-title / axis-title /
- * axis tick-label / legend font family readers exactly so a parsed
- * value slots straight back into the writer's emit path.
- */
-function parseDataLabelsFontFamily(dLbls: XmlElement): string | undefined {
-  const txPr = findChild(dLbls, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const latin = findChild(defRPr, "latin");
-  if (!latin) return undefined;
-  const raw = latin.attrs.typeface;
-  if (typeof raw !== "string") return undefined;
-  const trimmed = raw.trim();
-  if (trimmed.length === 0) return undefined;
-  return trimmed;
-}
-
-/**
- * Pull `<c:dLbls><c:spPr><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill></c:spPr></c:dLbls>` off a data-labels block. Returns
- * the 6-character uppercase hex string when the parser walks the full
- * chain and lands on an `<a:srgbClr val="RRGGBB"/>`. Theme references
- * (`<a:schemeClr>`), `<a:hslClr>`, `<a:sysClr>`, and `<a:prstClr>`
- * all collapse to `undefined` — only the literal RGB triple
- * round-trips losslessly through {@link writeChart}. Non-solid fills
- * (`<a:noFill>`, `<a:gradFill>`, `<a:pattFill>`, `<a:blipFill>`)
- * likewise drop to `undefined` so a round-trip never fabricates a fill
- * the writer cannot reproduce on emit. Malformed `val` tokens (wrong
- * length, non-hex characters) drop to `undefined` rather than fabricate
- * a value the writer would round-trip into a malformed `<a:srgbClr>`.
- *
- * The OOXML `<c:spPr>` block sits on `CT_DLbls` between `<c:numFmt>`
- * and `<c:txPr>` per the schema sequence (ECMA-376 Part 1,
- * §21.2.2.50). Distinct from {@link parseDataLabelsFontColor}: the
- * fill lives on `<c:dLbls><c:spPr>`, the font color lives on
- * `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr><a:solidFill>` — the two
- * readers walk disjoint paths so a caller can pin both knobs without
- * conflict. Mirrors the chart-title fill {@link parseTitleFillColor} /
- * axis-title fill {@link parseAxisTitleFillColor} / plot-area fill
- * {@link parsePlotAreaFillColor} / legend fill
- * {@link parseLegendFillColor} so a parsed value slots straight into
- * the writer-side {@link ChartDataLabels.fillColor}.
- */
-function parseDataLabelsFillColor(dLbls: XmlElement): string | undefined {
-  return parseSpPrFill(dLbls);
-}
-
-/**
- * Pull `<c:dLbls><c:spPr><a:ln><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill></a:ln></c:spPr></c:dLbls>` off a data-labels block.
- * Returns the data-labels border (line) color as a 6-character
- * uppercase hex string.
- *
- * The OOXML `<a:srgbClr>` element carries the literal sRGB color
- * (`CT_SRgbColor`, ECMA-376 Part 1, §20.1.2.3.32) inside the line's
- * solid fill choice (`CT_LineProperties`'s solid fill — §20.1.2.3.24).
- * The `<a:ln>` slot sits inside the `<c:spPr>` block on `<c:dLbls>`
- * alongside the optional `<a:solidFill>` fill child, in
- * `CT_ShapeProperties` schema order (fill before stroke). The
- * `<c:spPr>` itself sits between `<c:numFmt>` and `<c:txPr>` per
- * CT_DLbls (ECMA-376 Part 1, §21.2.2.50).
- *
- * The reader surfaces only the literal `<a:srgbClr>` form — absence,
- * non-solid line fills (`<a:noFill>` / `<a:gradFill>` / `<a:pattFill>`),
- * and theme-color references (`<a:schemeClr>`) all collapse to
- * `undefined` so a chart that pinned a stroke the writer cannot
- * reproduce on emit drops the field rather than fabricate one Excel
- * would render differently. Malformed `val` tokens (wrong length,
- * non-hex characters, alpha-channel forms, non-string escapes)
- * likewise drop to `undefined`.
- *
- * Mirrors the writer-side {@link ChartDataLabels.borderColor} so a
- * parsed value slots straight into {@link cloneChart} without
- * conversion. The lookup is scoped to direct children of `<c:dLbls>`
- * so a stray `<c:spPr>` elsewhere (e.g. on `<c:plotArea>` /
- * `<c:legend>` / `<c:title>` / a series) cannot leak in. Mirrors
- * {@link parseDataLabelsFillColor} — same `<c:spPr>` host element on
- * the same `<c:dLbls>` parent — but lands on the line
- * (`<a:ln><a:solidFill>`) child rather than the fill (`<a:solidFill>`)
- * child.
- */
-function parseDataLabelsBorderColor(dLbls: XmlElement): string | undefined {
-  return parseSpPrBorderColor(dLbls);
-}
-
-/**
- * Pull `<c:numFmt formatCode=".." sourceLinked=".."/>` off a
- * `<c:dLbls>` block. Returns `undefined` when the element is absent or
- * when `formatCode` is missing / empty (the OOXML schema requires the
- * attribute on every emitted `<c:numFmt>` so a fabricated empty record
- * cannot round-trip cleanly).
- *
- * `sourceLinked` accepts the same OOXML truthy / falsy spellings the
- * other boolean attributes do (`"1"` / `"true"` / `"0"` / `"false"`);
- * absence and the OOXML default `"0"` collapse to `undefined` so the
- * parsed shape stays minimal — only an explicit `"1"` / `"true"`
- * surfaces `true`. Mirrors how the axis-side numFmt parser shapes its
- * output.
- */
-function parseDataLabelsNumberFormat(el: XmlElement): ChartAxisNumberFormat | undefined {
-  const numFmt = findChild(el, "numFmt");
-  if (!numFmt) return undefined;
-  const formatCode = numFmt.attrs.formatCode;
-  if (typeof formatCode !== "string" || formatCode.length === 0) return undefined;
-  const out: ChartAxisNumberFormat = { formatCode };
-  const sourceLinked = numFmt.attrs.sourceLinked;
-  if (typeof sourceLinked === "string") {
-    if (sourceLinked === "1" || sourceLinked.toLowerCase() === "true") {
-      out.sourceLinked = true;
-    }
-  }
-  return out;
-}
-
 /**
  * Read a boolean-style `val` attribute. Excel emits `"1"` / `"0"` but
  * the OOXML spec also blesses `"true"` / `"false"`. Returns `undefined`
@@ -4424,431 +3864,6 @@ function parseDate1904(chartSpace: XmlElement): boolean | undefined {
 
 // ── Data Table ─────────────────────────────────────────────────────
 
-/**
- * Pull `<c:dTable>...</c:dTable>` off `<c:plotArea>`. Surfaces a
- * {@link ChartDataTable} whenever the source chart declares the
- * element; absence collapses to `undefined`.
- *
- * Each of the four boolean children (`<c:showHorzBorder>`,
- * `<c:showVertBorder>`, `<c:showOutline>`, `<c:showKeys>`) round-trips
- * literally — the reader does not collapse any per-field default
- * because all four are required on `CT_DTable` and Excel always emits
- * every one. Children that are missing or carry an unknown `val`
- * attribute drop to `undefined` rather than fabricate a flag the file
- * did not pin; the writer falls back to the OOXML reference defaults
- * (`true` for every child) on round-trip.
- */
-function parseDataTable(plotArea: XmlElement): ChartDataTable | undefined {
-  const el = findChild(plotArea, "dTable");
-  if (!el) return undefined;
-  const out: ChartDataTable = {};
-  const showHorzBorder = parseDataTableFlag(el, "showHorzBorder");
-  if (showHorzBorder !== undefined) out.showHorzBorder = showHorzBorder;
-  const showVertBorder = parseDataTableFlag(el, "showVertBorder");
-  if (showVertBorder !== undefined) out.showVertBorder = showVertBorder;
-  const showOutline = parseDataTableFlag(el, "showOutline");
-  if (showOutline !== undefined) out.showOutline = showOutline;
-  const showKeys = parseDataTableFlag(el, "showKeys");
-  if (showKeys !== undefined) out.showKeys = showKeys;
-  const fontSize = parseDataTableFontSize(el);
-  if (fontSize !== undefined) out.fontSize = fontSize;
-  const fontColor = parseDataTableFontColor(el);
-  if (fontColor !== undefined) out.fontColor = fontColor;
-  const bold = parseDataTableBold(el);
-  if (bold !== undefined) out.bold = bold;
-  const italic = parseDataTableItalic(el);
-  if (italic !== undefined) out.italic = italic;
-  const underline = parseDataTableUnderline(el);
-  if (underline !== undefined) out.underline = underline;
-  const strikethrough = parseDataTableStrikethrough(el);
-  if (strikethrough !== undefined) out.strikethrough = strikethrough;
-  const fontFamily = parseDataTableFontFamily(el);
-  if (fontFamily !== undefined) out.fontFamily = fontFamily;
-  const fillColor = parseDataTableFillColor(el);
-  if (fillColor !== undefined) out.fillColor = fillColor;
-  const borderColor = parseDataTableBorderColor(el);
-  if (borderColor !== undefined) out.borderColor = borderColor;
-  // `<c:dTable><c:spPr><a:ln w="EMU">` carries Excel's "Format Data
-  // Table -> Border -> Width" pin. Delegates to the shared
-  // {@link parseBorderWidthFromSpPr} so the EMU encoding and snap /
-  // clamp grammar match every other chart-frame border-width slot.
-  const borderWidth = parseBorderWidthFromSpPr(el);
-  if (borderWidth !== undefined) out.borderWidth = borderWidth;
-  // `<c:dTable><c:spPr><a:ln><a:prstDash val=".."/>` carries Excel's
-  // "Format Data Table -> Border -> Dash type" pin. Delegates to the
-  // shared {@link parseBorderDashFromSpPr} so the accept-or-drop
-  // grammar matches every chart-frame border-dash slot.
-  const borderDash = parseBorderDashFromSpPr(el);
-  if (borderDash !== undefined) out.borderDash = borderDash;
-  return out;
-}
-
-/**
- * Pull `<c:dTable><c:txPr><a:p><a:pPr><a:defRPr b=".."/></a:pPr></a:p>
- * </c:txPr></c:dTable>` off a data-table block. Returns the bold flag.
- *
- * The OOXML `b` attribute is the `xsd:boolean` bold flag on
- * `CT_TextCharacterProperties`. Only an explicit `b="1"` (or `"true"`)
- * surfaces `true`; the OOXML default `0` (and absence / malformed
- * tokens) collapses to `undefined` so absence and the default
- * round-trip identically through `cloneChart`. Mirrors the
- * chart-title / axis-title / axis tick-label / legend / data-label
- * bold readers exactly so a parsed value slots straight back into the
- * writer's emit path.
- */
-function parseDataTableBold(dTable: XmlElement): boolean | undefined {
-  const txPr = findChild(dTable, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.b;
-  if (typeof raw !== "string") return undefined;
-  switch (raw) {
-    case "1":
-    case "true":
-      return true;
-    case "0":
-    case "false":
-      return false;
-    default:
-      return undefined;
-  }
-}
-
-/**
- * Pull `<c:dTable><c:txPr><a:p><a:pPr><a:defRPr i=".."/></a:pPr></a:p>
- * </c:txPr></c:dTable>` off a data-table block. Returns the italic
- * flag.
- *
- * The OOXML `i` attribute is the `xsd:boolean` italic flag on
- * `CT_TextCharacterProperties`. An explicit `i="1"` / `"true"`
- * surfaces `true`; an explicit `i="0"` / `"false"` surfaces `false`
- * so a templated chart that pinned `i="0"` round-trips literally and
- * a clone target can override an upstream `i="1"` cleanly. Absence /
- * malformed tokens collapse to `undefined` so a fresh data table that
- * never pinned the flag round-trips identically through `cloneChart`.
- * Mirrors the data-table bold reader (and the chart-title /
- * axis-title / axis tick-label / legend / data-label italic readers
- * for the OOXML attribute layout) so a parsed value slots straight
- * back into the writer's emit path.
- */
-function parseDataTableItalic(dTable: XmlElement): boolean | undefined {
-  const txPr = findChild(dTable, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.i;
-  if (typeof raw !== "string") return undefined;
-  switch (raw) {
-    case "1":
-    case "true":
-      return true;
-    case "0":
-    case "false":
-      return false;
-    default:
-      return undefined;
-  }
-}
-
-/**
- * Pull `<c:dTable><c:txPr><a:p><a:pPr><a:defRPr u=".."/></a:pPr></a:p>
- * </c:txPr></c:dTable>` off a data-table block. Returns the underline
- * flag.
- *
- * The OOXML `u` attribute is the `ST_TextUnderlineType` enumeration
- * on `CT_TextCharacterProperties`. `u="sng"` (Excel's UI variant —
- * single underline) surfaces `true`; `u="none"` (the OOXML default)
- * surfaces `false` so a templated chart that pinned `u="none"`
- * round-trips literally and a clone target can override an upstream
- * `u="sng"` cleanly. The schema's other variants (`"dbl"`, `"heavy"`,
- * `"dotted"`, `"dotDash"`, `"wavy"`, etc.) and absence / malformed
- * tokens collapse to `undefined` so a fresh data table that never
- * pinned the flag round-trips identically through `cloneChart`. The
- * reader never silently downgrades a non-`"sng"` underline to a
- * single line on round-trip; only the boolean shape Excel's UI
- * exposes survives the parse.
- *
- * Mirrors the data-table bold reader pattern — `true` and `false`
- * round-trip explicitly so a clone can override either direction —
- * while the OOXML attribute layout matches the chart-title /
- * axis-title / axis tick-label / legend / data-label underline
- * readers exactly so a parsed value slots straight back into the
- * writer's emit path.
- */
-function parseDataTableUnderline(dTable: XmlElement): boolean | undefined {
-  const txPr = findChild(dTable, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.u;
-  if (raw === "sng") return true;
-  if (raw === "none") return false;
-  return undefined;
-}
-
-/**
- * Pull `<c:dTable><c:txPr><a:p><a:pPr><a:defRPr strike=".."/></a:pPr>
- * </a:p></c:txPr></c:dTable>` off a data-table block. Returns the
- * strikethrough flag.
- *
- * The OOXML `strike` attribute is the `ST_TextStrikeType`
- * enumeration on `CT_TextCharacterProperties`. Only
- * `strike="sngStrike"` (Excel's UI variant — single line) surfaces
- * `true`; the OOXML default `"noStrike"` and the non-UI variant
- * `"dblStrike"` (and any malformed token) collapse to `undefined` so
- * absence and `"noStrike"` round-trip identically through
- * `cloneChart`. Reporting `"dblStrike"` as `true` would silently
- * downgrade the choice to a single line on round-trip; the writer
- * emits only `"sngStrike"`, matching the boolean shape the UI
- * exposes. Mirrors the chart-title / axis-title / axis tick-label /
- * legend / data-label strikethrough readers exactly so a parsed value
- * slots straight back into the writer's emit path.
- */
-function parseDataTableStrikethrough(dTable: XmlElement): boolean | undefined {
-  const txPr = findChild(dTable, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.strike;
-  if (raw === "sngStrike") return true;
-  return undefined;
-}
-
-/**
- * Pull `<c:dTable><c:txPr><a:p><a:pPr><a:defRPr><a:latin
- * typeface=".."/></a:defRPr></a:pPr></a:p></c:txPr></c:dTable>` off
- * a data-table block. Returns the typeface string the table was
- * authored with.
- *
- * The OOXML `<a:latin>` element carries the typeface name on
- * `CT_TextFont` (ECMA-376 Part 1, §21.1.2.3.7). The reader trims
- * surrounding whitespace and reports the trimmed typeface; empty /
- * whitespace-only `typeface` attributes and missing `<a:latin>`
- * elements both collapse to `undefined` so absence and the empty
- * form round-trip identically through the writer. Non-string
- * `typeface` tokens (defensive — the XML parser only ever surfaces
- * strings) likewise drop to `undefined`.
- *
- * Returns `undefined` whenever the data-table block omits `<c:txPr>`
- * entirely or the canonical `<a:p><a:pPr><a:defRPr><a:latin>` chain
- * is malformed at any link. Mirrors the chart-title / axis-title /
- * axis tick-label / legend / data-label font family readers exactly
- * so a parsed value slots straight back into the writer's emit path.
- */
-function parseDataTableFontFamily(dTable: XmlElement): string | undefined {
-  const txPr = findChild(dTable, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const latin = findChild(defRPr, "latin");
-  if (!latin) return undefined;
-  const raw = latin.attrs.typeface;
-  if (typeof raw !== "string") return undefined;
-  const trimmed = raw.trim();
-  if (trimmed.length === 0) return undefined;
-  return trimmed;
-}
-
-/**
- * Pull `<c:dTable><c:spPr><a:solidFill><a:srgbClr val=".."/>
- * </a:solidFill></c:spPr></c:dTable>` off a data-table block.
- * Returns the data-table background fill color as a 6-character
- * uppercase hex string.
- *
- * The OOXML `<a:srgbClr>` element carries the literal sRGB color
- * (`CT_SRgbColor`, ECMA-376 Part 1, §20.1.2.3.32) inside the fill
- * choice of `<c:spPr>` (`CT_ShapeProperties`, §20.1.2.3.13). The
- * `<c:spPr>` slot sits inside `<c:dTable>` after the four required
- * boolean children (`<c:showHorzBorder>`, `<c:showVertBorder>`,
- * `<c:showOutline>`, `<c:showKeys>`) and before the optional
- * `<c:txPr>` per CT_DTable (ECMA-376 Part 1, §21.2.2.54).
- *
- * The reader surfaces only the literal `<a:srgbClr>` form — absence,
- * non-solid fills (`<a:noFill>` / `<a:gradFill>` / `<a:pattFill>` /
- * `<a:blipFill>`), and theme-color references (`<a:schemeClr>`) all
- * collapse to `undefined` so a chart that pinned a fill the writer
- * cannot reproduce on emit drops the field rather than fabricate one
- * Excel would render differently. Malformed `val` tokens (wrong
- * length, non-hex characters, alpha-channel forms, non-string escapes)
- * likewise drop to `undefined`.
- *
- * Mirrors the writer-side {@link ChartDataTable.fillColor} so a parsed
- * value slots straight into {@link cloneChart} without conversion.
- * The lookup is scoped to direct children of `<c:dTable>` so a stray
- * `<c:spPr>` elsewhere (e.g. on `<c:plotArea>` / `<c:legend>` /
- * `<c:title>` / a series) cannot leak into this field. Mirrors
- * {@link parsePlotAreaFillColor} / {@link parseLegendFillColor} —
- * same `<c:spPr><a:solidFill><a:srgbClr>` chain on a different host
- * element. Independent of {@link parseDataTableBorderColor}: the fill
- * lives on `<c:dTable><c:spPr><a:solidFill>`, the stroke lives on
- * `<c:dTable><c:spPr><a:ln><a:solidFill>` — the two readers walk
- * disjoint children of the same `<c:spPr>` block so a caller can pin
- * both knobs without conflict.
- */
-function parseDataTableFillColor(dTable: XmlElement): string | undefined {
-  return parseSpPrFill(dTable);
-}
-
-/**
- * Pull `<c:dTable><c:spPr><a:ln><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill></a:ln></c:spPr></c:dTable>` off a data-table block.
- * Returns the data-table border (line) color as a 6-character uppercase
- * hex string.
- *
- * The OOXML `<a:srgbClr>` element carries the literal sRGB color
- * (`CT_SRgbColor`, ECMA-376 Part 1, §20.1.2.3.32) inside the line's
- * solid fill choice (`CT_LineProperties`'s solid fill — §20.1.2.3.24).
- * The `<a:ln>` slot sits inside the `<c:spPr>` block on `<c:dTable>`
- * alongside the optional `<a:solidFill>` fill child, in
- * `CT_ShapeProperties` schema order (fill before stroke). The
- * `<c:spPr>` itself sits between the four required boolean children
- * (`<c:showHorzBorder>`, `<c:showVertBorder>`, `<c:showOutline>`,
- * `<c:showKeys>`) and the optional `<c:txPr>` per CT_DTable
- * (ECMA-376 Part 1, §21.2.2.54).
- *
- * The reader surfaces only the literal `<a:srgbClr>` form — absence,
- * non-solid line fills (`<a:noFill>` / `<a:gradFill>` / `<a:pattFill>`),
- * and theme-color references (`<a:schemeClr>`) all collapse to
- * `undefined` so a chart that pinned a stroke the writer cannot
- * reproduce on emit drops the field rather than fabricate one Excel
- * would render differently. Malformed `val` tokens (wrong length,
- * non-hex characters, alpha-channel forms, non-string escapes)
- * likewise drop to `undefined`.
- *
- * Mirrors the writer-side {@link ChartDataTable.borderColor} so a
- * parsed value slots straight into {@link cloneChart} without
- * conversion. The lookup is scoped to direct children of `<c:dTable>`
- * so a stray `<c:spPr>` elsewhere (e.g. on `<c:plotArea>` /
- * `<c:legend>` / `<c:title>` / a series) cannot leak in. Mirrors
- * {@link parseDataTableFillColor} — same `<c:spPr>` host element on
- * the same `<c:dTable>` parent — but lands on the line
- * (`<a:ln><a:solidFill>`) child rather than the fill (`<a:solidFill>`)
- * child.
- */
-function parseDataTableBorderColor(dTable: XmlElement): string | undefined {
-  return parseSpPrBorderColor(dTable);
-}
-
-/**
- * Pull `<c:dTable><c:txPr><a:p><a:pPr><a:defRPr><a:solidFill>
- * <a:srgbClr val="RRGGBB"/></a:solidFill></a:defRPr></a:pPr></a:p>
- * </c:txPr></c:dTable>` off a data-table block. Returns the
- * 6-character uppercase hex string when the parser walks the full
- * chain and lands on an `<a:srgbClr val="RRGGBB"/>`. Theme references
- * (`<a:schemeClr>`), `<a:hslClr>`, `<a:sysClr>`, and `<a:prstClr>`
- * all collapse to `undefined` — only the literal RGB triple round-
- * trips losslessly through {@link writeChart}. Malformed `val` tokens
- * (wrong length, non-hex characters) likewise drop to `undefined`
- * rather than fabricate a value the writer would round-trip into a
- * malformed `<a:srgbClr>`.
- *
- * Returns `undefined` whenever the data-table block omits `<c:txPr>`
- * entirely or the canonical chain is malformed at any link. Mirrors
- * the chart-title / axis-title / axis tick-label / legend / data-label
- * color readers exactly so a parsed value slots straight back into the
- * writer's emit path.
- */
-function parseDataTableFontColor(dTable: XmlElement): string | undefined {
-  const txPr = findChild(dTable, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const solidFill = findChild(defRPr, "solidFill");
-  if (!solidFill) return undefined;
-  const srgbClr = findChild(solidFill, "srgbClr");
-  if (!srgbClr) return undefined;
-  return normalizeRgbHex(srgbClr.attrs.val);
-}
-
-/**
- * Pull `<c:dTable><c:txPr><a:p><a:pPr><a:defRPr sz="N"/></a:pPr></a:p>
- * </c:txPr></c:dTable>` off a data-table block. Returns the font size
- * in points (`1..400`), or `undefined` when the element is absent /
- * the chain is malformed at any link / the surfaced value is out of
- * the supported range.
- *
- * The OOXML attribute is in 100ths of a point on
- * `CT_TextCharacterProperties`' `sz` slot (ECMA-376 Part 1,
- * §21.1.2.3.7); the reader divides by 100 at parse time so the
- * surfaced value matches what the user sees in Excel's UI. Mirrors
- * the chart-title / axis-title / axis tick-label / legend / data-label
- * font-size readers exactly so a parsed value slots straight back into
- * the writer's emit path.
- */
-function parseDataTableFontSize(dTable: XmlElement): number | undefined {
-  const txPr = findChild(dTable, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.sz;
-  if (typeof raw !== "string") return undefined;
-  const trimmed = raw.trim();
-  if (trimmed.length === 0) return undefined;
-  const parsed = Number.parseInt(trimmed, 10);
-  if (!Number.isFinite(parsed)) return undefined;
-  // Convert from 100ths of a point to points, rounding to the nearest
-  // 0.5pt to match the granularity Excel's UI exposes. Mirrors the
-  // chart-title / axis-title / tick-label / legend / data-label
-  // sibling parsers exactly so a parsed value flows through every
-  // typography slot without bookkeeping the units.
-  const halfSteps = Math.round((parsed / TITLE_FONT_SZ_PER_POINT) * 2);
-  const points = halfSteps / 2;
-  if (points < TITLE_FONT_SIZE_MIN_PT || points > TITLE_FONT_SIZE_MAX_PT) return undefined;
-  return points;
-}
-
-/**
- * Pull a single boolean child off `<c:dTable>`. Accepts the OOXML
- * truthy / falsy spellings (`"1"` / `"true"` / `"0"` / `"false"`);
- * unknown tokens, missing `val` attributes, and missing elements all
- * collapse to `undefined` rather than fabricate a flag the file did
- * not pin.
- */
-function parseDataTableFlag(dTable: XmlElement, local: string): boolean | undefined {
-  const el = findChild(dTable, local);
-  if (!el) return undefined;
-  const raw = el.attrs.val;
-  if (typeof raw !== "string") return undefined;
-  switch (raw) {
-    case "1":
-    case "true":
-      return true;
-    case "0":
-    case "false":
-      return false;
-    default:
-      return undefined;
-  }
-}
 
 // ── Protection ────────────────────────────────────────────────────
