+++ src/xlsx/chart-writer.ts
@@ -108,6 +108,11 @@ import {
   resolveTitleStrike,
   resolveTitleUnderline,
 } from "./chart/title";
+import { buildDataTable, resolveDataTable } from "./chart/dataTable";
+import {
+  buildChartLevelDataLabels,
+  buildSeriesDataLabels,
+} from "./chart/dataLabels";
 
 // ── Namespaces ───────────────────────────────────────────────────────
 
@@ -1089,488 +1094,6 @@ function normalizeChartSpaceBorderColor(value: string | undefined): string | und
 
 // ── Data Table ───────────────────────────────────────────────────────
 
-/**
- * Resolve the {@link SheetChart.dataTable} field into the four boolean
- * children `<c:dTable>` requires, or `undefined` to signal that the
- * writer should skip emission of the element.
- *
- * Returns `undefined` when:
- *  - The chart's family has no axes (pie / doughnut). The OOXML schema
- *    places `<c:dTable>` inside `<c:plotArea>` alongside the axes, so
- *    no axes means no slot to host the element.
- *  - The caller did not opt in (`dataTable` is `undefined` or `false`).
- *
- * Returns the four resolved booleans when the caller passed `true`
- * (every default `true`) or an object (per-field overrides on top of the
- * `true` defaults). Stray non-boolean inputs collapse to the matching
- * default rather than emit a token Excel rejects, mirroring how every
- * other chart-level boolean writer treats its input.
- */
-function resolveDataTable(chart: SheetChart):
-  | {
-      showHorzBorder: boolean;
-      showVertBorder: boolean;
-      showOutline: boolean;
-      showKeys: boolean;
-      fontSize: number | undefined;
-      fontColor: string | undefined;
-      bold: boolean | undefined;
-      italic: boolean | undefined;
-      underline: boolean | undefined;
-      strikethrough: boolean | undefined;
-      fontFamily: string | undefined;
-      fillColor: string | undefined;
-      borderColor: string | undefined;
-      borderWidth: number | undefined;
-      borderDash: ChartBorderDash | undefined;
-    }
-  | undefined {
-  // Pie / doughnut have no axes — the OOXML schema places `<c:dTable>`
-  // alongside `<c:catAx>` / `<c:valAx>`, so there is no slot for it on
-  // those families. Drop the field silently rather than emit an element
-  // Excel's strict validator would reject.
-  if (chart.type === "pie" || chart.type === "doughnut") return undefined;
-
-  const raw = chart.dataTable;
-  if (raw === undefined || raw === false) return undefined;
-
-  if (raw === true) {
-    return {
-      showHorzBorder: true,
-      showVertBorder: true,
-      showOutline: true,
-      showKeys: true,
-      fontSize: undefined,
-      fontColor: undefined,
-      bold: undefined,
-      italic: undefined,
-      underline: undefined,
-      strikethrough: undefined,
-      fontFamily: undefined,
-      fillColor: undefined,
-      borderColor: undefined,
-      borderWidth: undefined,
-      borderDash: undefined,
-    };
-  }
-
-  // Per-field overrides on top of the `true` defaults. Only literal
-  // `false` flips a flag — anything else (including stray `undefined`,
-  // `null`, or a non-boolean) falls back to the default `true` so the
-  // writer never emits a value the OOXML schema would refuse.
-  return {
-    showHorzBorder: raw.showHorzBorder !== false,
-    showVertBorder: raw.showVertBorder !== false,
-    showOutline: raw.showOutline !== false,
-    showKeys: raw.showKeys !== false,
-    fontSize: resolveDataTableFontSize(raw.fontSize),
-    fontColor: resolveDataTableFontColor(raw.fontColor),
-    bold: resolveDataTableBold(raw.bold),
-    italic: resolveDataTableItalic(raw.italic),
-    underline: resolveDataTableUnderline(raw.underline),
-    strikethrough: resolveDataTableStrikethrough(raw.strikethrough),
-    fontFamily: resolveDataTableFontFamily(raw.fontFamily),
-    fillColor: resolveDataTableFillColor(raw.fillColor),
-    borderColor: resolveDataTableBorderColor(raw.borderColor),
-    borderWidth: clampStrokeWidthPt(raw.borderWidth),
-    borderDash: normalizeBorderDash(raw.borderDash),
-  };
-}
-
-/**
- * Resolve `<c:dTable><c:txPr><a:p><a:pPr><a:defRPr sz="N"/></a:pPr></a:p>
- * </c:txPr></c:dTable>` from {@link ChartDataTable.fontSize}.
- *
- * Returns the size in points (`1..400`), or `undefined` when the caller
- * leaves the field unset / passed an out-of-range or non-numeric token.
- * Delegates to {@link normalizeTitleFontSize} so the chart-title /
- * axis-title / axis tick-label / legend / data-label / data-table
- * font-size resolvers share the same range, the same fractional rounding
- * (Excel's UI step is 0.5pt), and the same OOXML conversion (100ths of
- * a point at emit time).
- */
-function resolveDataTableFontSize(value: number | undefined): number | undefined {
-  return normalizeTitleFontSize(value);
-}
-
-/**
- * Resolve `<c:dTable><c:txPr><a:p><a:pPr><a:defRPr><a:solidFill>
- * <a:srgbClr val="RRGGBB"/></a:solidFill></a:defRPr></a:pPr></a:p>
- * </c:txPr></c:dTable>` from {@link ChartDataTable.fontColor}.
- *
- * Returns the 6-character uppercase hex string the writer emits, or
- * `undefined` when the caller leaves the field unset / passed a
- * malformed token. Delegates to {@link normalizeTitleColor} so the
- * accept-with-or-without-`#` grammar matches the chart-title /
- * axis-title / axis tick-label / legend / data-label color resolvers
- * exactly.
- */
-function resolveDataTableFontColor(value: string | undefined): string | undefined {
-  return normalizeTitleColor(value);
-}
-
-/**
- * Resolve `<c:dTable><c:txPr><a:p><a:pPr><a:defRPr b=".."/></a:pPr>
- * </a:p></c:txPr></c:dTable>` from {@link ChartDataTable.bold}.
- *
- * Returns the bold flag, or `undefined` when the caller leaves the
- * field unset / passed a non-boolean token. Mirrors the chart-title /
- * axis-title / axis tick-label / legend / data-label bold resolvers
- * exactly — only literal `true` / `false` pass through; non-boolean
- * tokens (typed escapes from an untyped caller) collapse to
- * `undefined`.
- */
-function resolveDataTableBold(value: boolean | undefined): boolean | undefined {
-  if (value === true) return true;
-  if (value === false) return false;
-  return undefined;
-}
-
-/**
- * Resolve `<c:dTable><c:txPr><a:p><a:pPr><a:defRPr i=".."/></a:pPr>
- * </a:p></c:txPr></c:dTable>` from {@link ChartDataTable.italic}.
- *
- * Returns the italic flag, or `undefined` when the caller leaves the
- * field unset / passed a non-boolean token. Mirrors the chart-title /
- * axis-title / axis tick-label / legend / data-label italic resolvers
- * (and the data-table bold resolver) exactly — only literal `true` /
- * `false` pass through; non-boolean tokens (typed escapes from an
- * untyped caller) collapse to `undefined`.
- */
-function resolveDataTableItalic(value: boolean | undefined): boolean | undefined {
-  if (value === true) return true;
-  if (value === false) return false;
-  return undefined;
-}
-
-/**
- * Resolve `<c:dTable><c:txPr><a:p><a:pPr><a:defRPr u=".."/></a:pPr>
- * </a:p></c:txPr></c:dTable>` from {@link ChartDataTable.underline}.
- *
- * Returns the underline flag, or `undefined` when the caller leaves
- * the field unset / passed a non-boolean token. Mirrors the
- * chart-title / axis-title / axis tick-label / legend / data-label
- * underline resolvers exactly — only literal `true` / `false` pass
- * through; non-boolean tokens (typed escapes from an untyped caller)
- * collapse to `undefined`. The writer translates `true` into
- * `u="sng"` (Excel's UI variant — single underline) and `false` into
- * `u="none"` at emit time.
- */
-function resolveDataTableUnderline(value: boolean | undefined): boolean | undefined {
-  if (value === true) return true;
-  if (value === false) return false;
-  return undefined;
-}
-
-/**
- * Resolve `<c:dTable><c:txPr><a:p><a:pPr><a:defRPr strike=".."/>
- * </a:pPr></a:p></c:txPr></c:dTable>` from
- * {@link ChartDataTable.strikethrough}.
- *
- * Returns `true` when the caller pins the strikethrough flag literally;
- * every other value (explicit `false`, absence, non-boolean tokens
- * leaking past the type guard) collapses to `undefined` so the writer
- * never emits a `strike` attribute below `"sngStrike"`. The OOXML
- * default `"noStrike"` is functionally identical to absence — the
- * writer keeps the surfaced shape consistent with what Excel's UI
- * authors (`"sngStrike"` only, never `"noStrike"` or `"dblStrike"`),
- * mirroring how `resolveTitleStrike` / `resolveLegendStrikethrough` /
- * `resolveDataLabelsStrikethrough` land on their `<a:defRPr>` slots.
- */
-function resolveDataTableStrikethrough(value: boolean | undefined): boolean | undefined {
-  if (value === true) return true;
-  return undefined;
-}
-
-/**
- * Resolve `<c:dTable><c:txPr><a:p><a:pPr><a:defRPr><a:latin
- * typeface=".."/></a:defRPr></a:pPr></a:p></c:txPr></c:dTable>` from
- * {@link ChartDataTable.fontFamily}.
- *
- * Returns the trimmed typeface string the writer emits, or
- * `undefined` when the caller leaves the field unset / passed an
- * empty / whitespace-only / non-string token. Mirrors the chart-title
- * / axis-title / axis tick-label / legend / data-label font family
- * resolvers exactly so a single configuration call threads cleanly
- * through every typography slot Excel exposes.
- */
-function resolveDataTableFontFamily(value: string | undefined): string | undefined {
-  if (typeof value !== "string") return undefined;
-  const trimmed = value.trim();
-  if (trimmed.length === 0) return undefined;
-  return trimmed;
-}
-
-/**
- * Resolve `<c:dTable><c:spPr><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill></c:spPr></c:dTable>` from
- * {@link ChartDataTable.fillColor}.
- *
- * Returns the 6-character uppercase hex string the writer emits, or
- * `undefined` when the caller leaves the field unset / passed a
- * malformed token. Delegates to {@link normalizeTitleColor} so the
- * accept-with-or-without-`#` grammar matches the chart-title /
- * plot-area / legend / chart-space / axis-title fill color resolvers
- * exactly. The `<c:spPr>` slot lives between the four required
- * boolean children and the optional `<c:txPr>` per CT_DTable
- * (ECMA-376 Part 1, §21.2.2.54), distinct from the `<c:txPr>` block
- * that carries {@link ChartDataTable.fontColor}.
- */
-function resolveDataTableFillColor(value: string | undefined): string | undefined {
-  return normalizeTitleColor(value);
-}
-
-/**
- * Resolve `<c:dTable><c:spPr><a:ln><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill></a:ln></c:spPr></c:dTable>` from
- * {@link ChartDataTable.borderColor}.
- *
- * Returns the 6-character uppercase hex string the writer emits, or
- * `undefined` when the caller leaves the field unset / passed a
- * malformed token. Delegates to {@link normalizeTitleColor} so the
- * accept-with-or-without-`#` grammar matches every other `<a:srgbClr>`
- * fill / line slot exactly. Composes independently with
- * {@link ChartDataTable.fillColor} — the two knobs share the same
- * `<c:spPr>` host on `<c:dTable>` but land on different children
- * (`<a:solidFill>` for the fill, `<a:ln><a:solidFill>` for the stroke).
- * Mirrors the chart-title / axis-title / chart-space / plot-area /
- * legend `<c:spPr>` border slots — same hex grammar, same `<a:ln>`
- * slot on the `CT_ShapeProperties` schema — but lands on `<c:dTable>`'s
- * own `<c:spPr>` block.
- */
-function resolveDataTableBorderColor(value: string | undefined): string | undefined {
-  return normalizeTitleColor(value);
-}
-
-/**
- * Serialize a resolved data-table into `<c:dTable>` with its four
- * required boolean children, in the order CT_DTable mandates:
- * `showHorzBorder`, `showVertBorder`, `showOutline`, `showKeys`. When
- * a typography knob is pinned the writer emits an additional `<c:txPr>`
- * block after the four booleans; absence collapses to the existing
- * four-child shape so a fresh chart with no typography pin matches
- * Excel's reference serialization byte-for-byte.
- *
- * The writer always emits all four boolean children — the OOXML schema
- * marks them required on `CT_DTable`, and Excel's reference
- * serialization includes every one even when the caller leaves it at
- * the default. The optional `<c:spPr>` / `<c:extLst>` children are
- * skipped because hucre's data-table model does not surface fill /
- * extension styling yet.
- */
-function buildDataTable(table: {
-  showHorzBorder: boolean;
-  showVertBorder: boolean;
-  showOutline: boolean;
-  showKeys: boolean;
-  fontSize: number | undefined;
-  fontColor: string | undefined;
-  bold: boolean | undefined;
-  italic: boolean | undefined;
-  underline: boolean | undefined;
-  strikethrough: boolean | undefined;
-  fontFamily: string | undefined;
-  fillColor: string | undefined;
-  borderColor: string | undefined;
-  borderWidth: number | undefined;
-  borderDash: ChartBorderDash | undefined;
-}): string {
-  const children: string[] = [
-    xmlSelfClose("c:showHorzBorder", { val: table.showHorzBorder ? 1 : 0 }),
-    xmlSelfClose("c:showVertBorder", { val: table.showVertBorder ? 1 : 0 }),
-    xmlSelfClose("c:showOutline", { val: table.showOutline ? 1 : 0 }),
-    xmlSelfClose("c:showKeys", { val: table.showKeys ? 1 : 0 }),
-  ];
-  // CT_DTable schema places `<c:spPr>` after the four required
-  // boolean children, before `<c:txPr>` and the optional `<c:extLst>`
-  // (ECMA-376 Part 1, §21.2.2.54). The writer skips emission entirely
-  // when no fill / border knob is pinned so a fresh chart matches
-  // Excel's reference serialization byte-for-byte.
-  const spPrXml = buildDataTableSpPr(
-    table.fillColor,
-    table.borderColor,
-    table.borderWidth,
-    table.borderDash,
-  );
-  if (spPrXml !== undefined) children.push(spPrXml);
-  // CT_DTable schema places `<c:txPr>` after `<c:spPr>` and before
-  // the optional `<c:extLst>` (ECMA-376 Part 1, §21.2.2.54). The writer
-  // skips emission entirely when no typography knob is pinned so a
-  // fresh chart matches Excel's reference serialization byte-for-byte.
-  const txPrXml = buildDataTableTxPr(
-    table.fontSize,
-    table.fontColor,
-    table.bold,
-    table.italic,
-    table.underline,
-    table.strikethrough,
-    table.fontFamily,
-  );
-  if (txPrXml !== undefined) children.push(txPrXml);
-  return xmlElement("c:dTable", undefined, children);
-}
-
-/**
- * Build the optional `<c:spPr>` block inside `<c:dTable>`. Surfaces
- * the solid fill color knob ({@link ChartDataTable.fillColor}) and
- * the border (line) color knob ({@link ChartDataTable.borderColor}) —
- * every other `<c:spPr>` child (`<a:effectLst>` effects, gradient /
- * pattern / picture fills, line dash / width / compound styles) is
- * intentionally not modelled at this layer.
- *
- * Returns `undefined` when both fields are unset / malformed so the
- * writer skips the entire `<c:spPr>` block — an empty `<c:spPr/>`
- * collapses to the inherited theme fill / stroke Excel picks anyway,
- * and omitting it keeps untouched chart XML byte-clean. When at least
- * one knob lands on the wire, the children are emitted in
- * `CT_ShapeProperties` schema order: `<a:solidFill>` (fill) then
- * `<a:ln>` (line / stroke).
- *
- * Mirrors {@link buildPlotAreaSpPr} / {@link buildChartSpaceSpPr}
- * but on a distinct host element — the data-table fill / stroke
- * paint the background and outline of the table grid, while the
- * plot-area / chart-space variants paint the inner band / entire
- * chart frame.
- */
-function buildDataTableSpPr(
-  fillColor: string | undefined,
-  borderColor: string | undefined,
-  borderWidthPt: number | undefined,
-  borderDash: ChartBorderDash | undefined,
-): string | undefined {
-  if (
-    fillColor === undefined &&
-    borderColor === undefined &&
-    borderWidthPt === undefined &&
-    borderDash === undefined
-  ) {
-    return undefined;
-  }
-  const children: string[] = [];
-  if (fillColor !== undefined) {
-    children.push(
-      xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: fillColor })]),
-    );
-  }
-  if (borderColor !== undefined || borderWidthPt !== undefined || borderDash !== undefined) {
-    const lnAttrs: Record<string, string | number> = {};
-    if (borderWidthPt !== undefined) {
-      // OOXML stores stroke width in EMU (1 pt = 12 700 EMU). Round to
-      // the nearest integer because the schema types `w` as `xsd:int`.
-      lnAttrs.w = Math.round(borderWidthPt * EMU_PER_PT);
-    }
-    const lnChildren: string[] = [];
-    if (borderColor !== undefined) {
-      lnChildren.push(
-        xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: borderColor })]),
-      );
-    }
-    // `<a:prstDash>` follows `<a:solidFill>` per CT_LineProperties
-    // schema sequence (ECMA-376 Part 1, §20.1.2.3.24).
-    if (borderDash !== undefined) {
-      lnChildren.push(xmlSelfClose("a:prstDash", { val: borderDash }));
-    }
-    children.push(
-      lnChildren.length === 0
-        ? xmlSelfClose("a:ln", lnAttrs)
-        : xmlElement("a:ln", Object.keys(lnAttrs).length > 0 ? lnAttrs : undefined, lnChildren),
-    );
-  }
-  return xmlElement("c:spPr", undefined, children);
-}
-
-/**
- * Build the `<c:txPr>` block that carries a data-table's typography
- * pins (currently font size and font color). Returns `undefined` when
- * every input is unset so the caller can elide the element entirely
- * (Excel's reference serialization omits `<c:txPr>` from `<c:dTable>`
- * when the table renders at the theme-default style).
- *
- * The emitted block mirrors the minimal `<c:txPr>` shape Excel writes
- * when the user pins a data-table typography knob — `<a:bodyPr/>`,
- * `<a:lstStyle/>`, and the `<a:p><a:pPr><a:defRPr sz="N"/></a:pPr>
- * <a:endParaRPr/></a:p>` paragraph stub Excel always emits hosts the
- * typography attributes on `<a:defRPr>`. The `<a:defRPr>` element
- * expands from self-closing to wrapping a single `<a:solidFill>` child
- * when a color is set; otherwise the writer keeps the existing
- * self-closing form so a fresh chart with no custom color matches
- * Excel's reference serialization byte-for-byte. Mirrors the
- * chart-title / axis-title / tick-label / legend / data-label
- * `<c:txPr>` slots exactly so a re-parse picks the values off the
- * canonical default-paragraph slot every other typography reader
- * expects.
- */
-function buildDataTableTxPr(
-  fontSizePt: number | undefined,
-  rgbHex: string | undefined,
-  bold: boolean | undefined,
-  italic: boolean | undefined,
-  underline: boolean | undefined,
-  strikethrough: boolean | undefined,
-  fontFamily: string | undefined,
-): string | undefined {
-  if (
-    fontSizePt === undefined &&
-    rgbHex === undefined &&
-    bold === undefined &&
-    italic === undefined &&
-    underline === undefined &&
-    strikethrough === undefined &&
-    fontFamily === undefined
-  )
-    return undefined;
-  const defRPrAttrs: Record<string, string | number> = {};
-  if (fontSizePt !== undefined) defRPrAttrs.sz = fontSizePt * TITLE_FONT_SZ_PER_POINT;
-  if (bold !== undefined) defRPrAttrs.b = bold ? 1 : 0;
-  if (italic !== undefined) defRPrAttrs.i = italic ? 1 : 0;
-  if (underline !== undefined) defRPrAttrs.u = underline ? "sng" : "none";
-  // Strikethrough rides as `strike="sngStrike"` on the same
-  // `<a:defRPr>` slot. Absence collapses to omitting the attribute
-  // entirely (the OOXML default `"noStrike"` is functionally
-  // identical to absence — the reader collapses both to `undefined`).
-  // The writer never emits `"noStrike"` or `"dblStrike"` so the
-  // surfaced shape stays consistent with Excel's UI checkbox.
-  if (strikethrough === true) defRPrAttrs.strike = "sngStrike";
-  // OOXML's `<a:defRPr><a:solidFill><a:srgbClr val="RRGGBB"/>
-  // </a:solidFill></a:defRPr>` carries the data-table font color.
-  // Absence (`undefined`) collapses to skipping the `<a:solidFill>`
-  // child entirely so the data table inherits the theme text color
-  // (Excel's reference behavior for fresh data tables that have not
-  // had a custom color picked).
-  const solidFillChild = rgbHex
-    ? xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: rgbHex })])
-    : undefined;
-  // OOXML's `<a:defRPr><a:latin typeface=".."/></a:defRPr>` carries
-  // the data-table font family. The `<a:latin>` element follows
-  // `<a:solidFill>` per the CT_TextCharacterProperties child sequence
-  // (ECMA-376 Part 1, §21.1.2.3.7). Absence (`undefined`) collapses
-  // to omitting the entire `<a:latin>` element so the data table
-  // inherits the theme typeface (Excel's reference behavior for
-  // fresh data tables that have not had a custom font picked).
-  const latinChild = fontFamily ? xmlSelfClose("a:latin", { typeface: fontFamily }) : undefined;
-  // When a fill color or a typeface is set the `<a:defRPr>` slot
-  // expands from self-closing to wrapping the children; otherwise the
-  // writer keeps the existing self-closing form so a fresh chart with
-  // no custom color or font matches Excel's reference serialization
-  // byte-for-byte. Children are emitted in
-  // CT_TextCharacterProperties' canonical schema order: solidFill
-  // first, then latin.
-  const defRPrChildren: string[] = [];
-  if (solidFillChild) defRPrChildren.push(solidFillChild);
-  if (latinChild) defRPrChildren.push(latinChild);
-  const defRPr =
-    defRPrChildren.length > 0
-      ? xmlElement("a:defRPr", defRPrAttrs, defRPrChildren)
-      : xmlSelfClose("a:defRPr", defRPrAttrs);
-  return xmlElement("c:txPr", undefined, [
-    xmlSelfClose("a:bodyPr"),
-    xmlSelfClose("a:lstStyle"),
-    xmlElement("a:p", undefined, [
-      xmlElement("a:pPr", undefined, [defRPr]),
-      xmlSelfClose("a:endParaRPr", { lang: "en-US" }),
-    ]),
-  ]);
-}
 
 // ── Protection ───────────────────────────────────────────────────────
 
@@ -4467,546 +3990,6 @@ function normalizeRgbHex(value: string | undefined): string | undefined {
 
 // ── Data Labels ──────────────────────────────────────────────────────
 
-/**
- * Resolve and emit the `<c:dLbls>` element for a single series.
- *
- * Series override semantics:
- *
- * - Series sets `dataLabels: false`  → emit a `delete=1` block to
- *   suppress this series even when the chart-level default enables labels.
- * - Series sets `dataLabels: <obj>`  → emit `<obj>`. Chart-level config is ignored.
- * - Series omits `dataLabels`        → no per-series `<c:dLbls>`. Excel
- *   inherits the chart-type-level `<c:dLbls>` block emitted by
- *   `buildChartLevelDataLabels` instead.
- *
- * Returns `undefined` when nothing should be emitted at the series level.
- */
-function buildSeriesDataLabels(
-  seriesDLbls: ChartDataLabels | false | undefined,
-  chartDLbls: ChartDataLabels | undefined,
-  chartType: WriteChartKind,
-): string | undefined {
-  if (seriesDLbls === false) {
-    // Suppress this series even when chart-level labels are on.
-    return xmlElement("c:dLbls", undefined, [
-      xmlElement("c:dLbl", undefined, [
-        xmlSelfClose("c:idx", { val: 0 }),
-        xmlSelfClose("c:delete", { val: 1 }),
-      ]),
-      xmlSelfClose("c:delete", { val: 1 }),
-    ]);
-  }
-  if (seriesDLbls) {
-    return buildDataLabelsBody(seriesDLbls, chartType);
-  }
-  // Series doesn't override → fall through to chart-level. Returning
-  // undefined here keeps the chart-level <c:dLbls> as the single source
-  // of truth so we don't duplicate the same toggles N times.
-  void chartDLbls;
-  return undefined;
-}
-
-/**
- * Build the chart-type-level `<c:dLbls>` block from
- * {@link SheetChart.dataLabels}. Returns `undefined` when no chart-level
- * labels are configured.
- */
-function buildChartLevelDataLabels(chart: SheetChart): string | undefined {
-  if (!chart.dataLabels) return undefined;
-  return buildDataLabelsBody(chart.dataLabels, chart.type);
-}
-
-/**
- * Render the OOXML `<c:dLbls>` body. Element order follows CT_DLbls:
- * delete? before numFmt? before spPr? before txPr? before dLblPos? before
- * showLegendKey, showVal, showCatName, showSerName, showPercent,
- * showBubbleSize, separator?, showLeaderLines? — toggles must appear
- * in that exact order or Excel ignores the block.
- *
- * `chartType` lets the builder gate the pie / doughnut-only
- * `<c:showLeaderLines>` element — the OOXML schema scopes that flag to
- * `EG_DLbls` for `CT_PieChart` / `CT_DoughnutChart` only, so the writer
- * silently drops `dl.showLeaderLines` on every other family rather than
- * emit a child Excel's strict validator would reject.
- */
-function buildDataLabelsBody(dl: ChartDataLabels, chartType: WriteChartKind): string {
-  const children: string[] = [];
-
-  // `<c:numFmt>` sits at the head of the CT_DLbls sequence (before
-  // `<c:spPr>` / `<c:txPr>` / `<c:dLblPos>` / the show* toggles). The
-  // writer skips emission entirely when the caller leaves `numberFormat`
-  // unset so a fresh chart matches Excel's reference shape (no number
-  // override means Excel inherits from the source cells).
-  const numFmt = resolveDataLabelsNumberFormat(dl.numberFormat);
-  if (numFmt) {
-    children.push(
-      xmlSelfClose("c:numFmt", {
-        formatCode: numFmt.formatCode,
-        sourceLinked: numFmt.sourceLinked === true ? 1 : 0,
-      }),
-    );
-  }
-
-  // `<c:spPr>` sits between `<c:numFmt>` and `<c:txPr>` in the
-  // CT_DLbls schema sequence (ECMA-376 Part 1, §21.2.2.50). The writer
-  // authors `<a:solidFill>` here from {@link ChartDataLabels.fillColor}
-  // and `<a:ln>` here from {@link ChartDataLabels.borderColor}; other
-  // `CT_ShapeProperties` children (gradient / pattern fills, line dash
-  // / width / compound styles, `<a:effectLst>` effects) are not
-  // modelled at this layer. The builder skips emission entirely when
-  // no fill / border is pinned so a fresh chart matches Excel's
-  // reference shape (no `<c:spPr>` block on a theme-default data-
-  // labels rendering — typically a transparent label background with
-  // no border). When at least one knob lands on the wire, the children
-  // are emitted in `CT_ShapeProperties` schema order: `<a:solidFill>`
-  // (fill) then `<a:ln>` (line / stroke). Distinct from the
-  // `<a:defRPr><a:solidFill>` font-color slot inside `<c:txPr>` that
-  // {@link ChartDataLabels.fontColor} pins — the three knobs target
-  // different children of `<c:dLbls>` so a caller can pin them all
-  // without conflict.
-  const spPrXml = buildDataLabelsSpPr(
-    resolveDataLabelsFillColor(dl.fillColor),
-    resolveDataLabelsBorderColor(dl.borderColor),
-    clampStrokeWidthPt(dl.borderWidth),
-    normalizeBorderDash(dl.borderDash),
-  );
-  if (spPrXml !== undefined) {
-    children.push(spPrXml);
-  }
-
-  // CT_DLbls schema places `<c:txPr>` between `<c:spPr>` and
-  // `<c:dLblPos>` (ECMA-376 Part 1, §21.2.2.50). The block currently
-  // carries the data-label font size, font color, bold flag, italic
-  // flag, underline flag, strikethrough flag, and font family — every
-  // typography pin lands on the same `<a:defRPr>` slot. The writer
-  // skips the entire block when no font knob is pinned so a fresh
-  // chart matches Excel's reference shape.
-  const txPrXml = buildDataLabelsTxPr(
-    resolveDataLabelsFontSize(dl.fontSize),
-    resolveDataLabelsFontColor(dl.fontColor),
-    resolveDataLabelsBold(dl.bold),
-    resolveDataLabelsItalic(dl.italic),
-    resolveDataLabelsUnderline(dl.underline),
-    resolveDataLabelsStrikethrough(dl.strikethrough),
-    resolveDataLabelsFontFamily(dl.fontFamily),
-  );
-  if (txPrXml !== undefined) {
-    children.push(txPrXml);
-  }
-
-  if (dl.position) {
-    children.push(xmlSelfClose("c:dLblPos", { val: dl.position }));
-  }
-
-  // OOXML requires showLegendKey to appear first when any toggle is set.
-  // Always emit it explicitly so the rendered XML is deterministic.
-  // Non-boolean inputs collapse to `false` to keep the on-the-wire output
-  // stable, mirroring how the other `show*` toggles treat their inputs.
-  children.push(xmlSelfClose("c:showLegendKey", { val: dl.showLegendKey === true ? 1 : 0 }));
-  children.push(xmlSelfClose("c:showVal", { val: dl.showValue ? 1 : 0 }));
-  children.push(xmlSelfClose("c:showCatName", { val: dl.showCategoryName ? 1 : 0 }));
-  children.push(xmlSelfClose("c:showSerName", { val: dl.showSeriesName ? 1 : 0 }));
-  children.push(xmlSelfClose("c:showPercent", { val: dl.showPercent ? 1 : 0 }));
-  children.push(xmlSelfClose("c:showBubbleSize", { val: 0 }));
-
-  if (dl.separator !== undefined) {
-    children.push(xmlElement("c:separator", undefined, xmlEscape(dl.separator)));
-  }
-
-  // `<c:showLeaderLines>` sits at the tail of the `EG_DLbls` group
-  // (after `<c:separator>`, before `<c:extLst>`). The OOXML schema
-  // scopes the element to pie / doughnut chart families exclusively
-  // (`EG_DLbls` for `CT_PieChart` / `CT_DoughnutChart` only — bar /
-  // column / line / area / scatter route through `EG_DLblsShared` which
-  // omits it). The writer drops the field silently on every non-pie /
-  // non-doughnut family to mirror Excel's reference serialization.
-  //
-  // The OOXML default is `true` (Excel paints leader lines on every
-  // label that gets pushed outside its slice). Only an explicit
-  // `false` flips the toggle, so absence and the default round-trip
-  // identically through {@link parseChart}. Non-boolean inputs collapse
-  // to the default, mirroring how the other `show*` toggles treat
-  // their inputs.
-  if ((chartType === "pie" || chartType === "doughnut") && dl.showLeaderLines === false) {
-    children.push(xmlSelfClose("c:showLeaderLines", { val: 0 }));
-  }
-
-  return xmlElement("c:dLbls", undefined, children);
-}
-
-/**
- * Resolve the `<c:numFmt>` value emitted inside `<c:dLbls>`.
- *
- * Returns `undefined` when the caller leaves `numberFormat` unset or
- * when `formatCode` is missing / non-string / empty — the OOXML schema
- * requires `formatCode` on every emitted `<c:numFmt>` so a malformed
- * record is dropped rather than fabricate a placeholder Excel rejects.
- * Mirrors how the axis-side number-format pipeline shapes its output
- * so a parsed value can flow straight from the read side back into the
- * write side without transformation.
- *
- * `sourceLinked` is normalized to a literal boolean — only `true`
- * survives, every other shape (`undefined` / `false` / non-boolean)
- * collapses so the writer's `val` attribute defaults to `0`.
- */
-function resolveDataLabelsNumberFormat(
-  value: ChartAxisNumberFormat | undefined,
-): ChartAxisNumberFormat | undefined {
-  if (!value) return undefined;
-  const formatCode = value.formatCode;
-  if (typeof formatCode !== "string" || formatCode.length === 0) return undefined;
-  const out: ChartAxisNumberFormat = { formatCode };
-  if (value.sourceLinked === true) out.sourceLinked = true;
-  return out;
-}
-
-/**
- * Resolve `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr sz="N"/></a:pPr>
- * </a:p></c:txPr></c:dLbls>` from {@link ChartDataLabels.fontSize}.
- *
- * Returns the size in points (`1..400`), or `undefined` when the
- * caller leaves the field unset / passed an out-of-range or
- * non-numeric token. Delegates to {@link normalizeTitleFontSize} so
- * the chart-title / axis-title / axis tick-label / legend / data-label
- * font-size resolvers share the same range, the same fractional
- * rounding (Excel's UI step is 0.5pt), and the same OOXML conversion
- * (100ths of a point at emit time).
- */
-function resolveDataLabelsFontSize(value: number | undefined): number | undefined {
-  return normalizeTitleFontSize(value);
-}
-
-/**
- * Resolve `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr><a:solidFill>
- * <a:srgbClr val="RRGGBB"/></a:solidFill></a:defRPr></a:pPr></a:p>
- * </c:txPr></c:dLbls>` from {@link ChartDataLabels.fontColor}.
- *
- * Returns the 6-character uppercase hex string the writer emits, or
- * `undefined` when the caller leaves the field unset / passed a
- * malformed token. Delegates to {@link normalizeTitleColor} so the
- * accept-with-or-without-`#` grammar matches the chart-title /
- * axis-title / axis tick-label / legend color resolvers exactly.
- */
-function resolveDataLabelsFontColor(value: string | undefined): string | undefined {
-  return normalizeTitleColor(value);
-}
-
-/**
- * Resolve `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr b=".."/></a:pPr>
- * </a:p></c:txPr></c:dLbls>` from {@link ChartDataLabels.bold}.
- *
- * Returns the bold flag, or `undefined` when the caller leaves the
- * field unset / passed a non-boolean token. Mirrors the chart-title /
- * axis-title / axis tick-label / legend bold resolvers exactly —
- * only literal `true` / `false` pass through; non-boolean tokens
- * (typed escapes from an untyped caller) collapse to `undefined`.
- */
-function resolveDataLabelsBold(value: boolean | undefined): boolean | undefined {
-  if (value === true) return true;
-  if (value === false) return false;
-  return undefined;
-}
-
-/**
- * Resolve `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr i=".."/></a:pPr>
- * </a:p></c:txPr></c:dLbls>` from {@link ChartDataLabels.italic}.
- *
- * Returns the italic flag, or `undefined` when the caller leaves the
- * field unset / passed a non-boolean token. Mirrors the chart-title /
- * axis-title / axis tick-label / legend italic resolvers exactly —
- * only literal `true` / `false` pass through; non-boolean tokens
- * (typed escapes from an untyped caller) collapse to `undefined`.
- */
-function resolveDataLabelsItalic(value: boolean | undefined): boolean | undefined {
-  if (value === true) return true;
-  if (value === false) return false;
-  return undefined;
-}
-
-/**
- * Resolve `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr u=".."/></a:pPr>
- * </a:p></c:txPr></c:dLbls>` from {@link ChartDataLabels.underline}.
- *
- * Returns the underline flag, or `undefined` when the caller leaves
- * the field unset / passed a non-boolean token. Mirrors the
- * chart-title / axis-title / axis tick-label / legend underline
- * resolvers exactly — only literal `true` / `false` pass through;
- * non-boolean tokens (typed escapes from an untyped caller) collapse
- * to `undefined`. The writer translates `true` into `u="sng"` (Excel's
- * UI variant — single underline) and `false` into `u="none"` at emit
- * time.
- */
-function resolveDataLabelsUnderline(value: boolean | undefined): boolean | undefined {
-  if (value === true) return true;
-  if (value === false) return false;
-  return undefined;
-}
-
-/**
- * Resolve `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr strike=".."/>
- * </a:pPr></a:p></c:txPr></c:dLbls>` from
- * {@link ChartDataLabels.strikethrough}.
- *
- * Returns `true` when the caller pins the strikethrough flag
- * literally; every other value (explicit `false`, absence, non-boolean
- * tokens leaking past the type guard) collapses to `undefined` so the
- * writer never emits a `strike` attribute below `"sngStrike"`. The
- * OOXML default `"noStrike"` is functionally identical to absence —
- * the writer keeps the surfaced shape consistent with what Excel's
- * UI authors (`"sngStrike"` only, never `"noStrike"` or
- * `"dblStrike"`), mirroring how `resolveTitleStrike` /
- * `resolveLegendStrikethrough` land on their `<a:defRPr>` slots.
- */
-function resolveDataLabelsStrikethrough(value: boolean | undefined): boolean | undefined {
-  if (value === true) return true;
-  return undefined;
-}
-
-/**
- * Resolve `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr><a:latin
- * typeface=".."/></a:defRPr></a:pPr></a:p></c:txPr></c:dLbls>` from
- * {@link ChartDataLabels.fontFamily}.
- *
- * Returns the trimmed typeface string the writer emits, or
- * `undefined` when the caller leaves the field unset / passed an
- * empty / whitespace-only / non-string token. Mirrors the chart-title
- * / axis-title / axis tick-label / legend font family resolvers
- * exactly so a single configuration call threads cleanly through
- * every typography slot Excel exposes.
- */
-function resolveDataLabelsFontFamily(value: string | undefined): string | undefined {
-  if (typeof value !== "string") return undefined;
-  const trimmed = value.trim();
-  if (trimmed.length === 0) return undefined;
-  return trimmed;
-}
-
-/**
- * Resolve `<c:dLbls><c:spPr><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill></c:spPr></c:dLbls>` from {@link ChartDataLabels.fillColor}.
- *
- * Returns the 6-character uppercase hex string the writer emits, or
- * `undefined` when the caller leaves the field unset / passed a
- * malformed token. Delegates to {@link normalizeTitleColor} so the
- * accept-with-or-without-`#` grammar matches the chart-title /
- * axis-title / plot-area / legend / data-label color resolvers
- * exactly. Distinct from {@link resolveDataLabelsFontColor}: the fill
- * lives on `<c:dLbls><c:spPr>`, the font color lives on
- * `<c:dLbls><c:txPr><a:p><a:pPr><a:defRPr><a:solidFill>` — the two
- * resolvers feed disjoint slots so a caller can pin both without
- * conflict.
- */
-function resolveDataLabelsFillColor(value: string | undefined): string | undefined {
-  return normalizeTitleColor(value);
-}
-
-/**
- * Resolve `<c:dLbls><c:spPr><a:ln><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill></a:ln></c:spPr></c:dLbls>` from
- * {@link ChartDataLabels.borderColor}.
- *
- * Returns the 6-character uppercase hex string the writer emits, or
- * `undefined` when the caller leaves the field unset / passed a
- * malformed token. Delegates to {@link normalizeTitleColor} so the
- * accept-with-or-without-`#` grammar matches every other `<a:srgbClr>`
- * fill / line slot exactly. Composes independently with
- * {@link ChartDataLabels.fillColor} — the two knobs share the same
- * `<c:spPr>` host on `<c:dLbls>` but land on different children
- * (`<a:solidFill>` for the fill, `<a:ln><a:solidFill>` for the
- * stroke). Mirrors the chart-title / axis-title / chart-space /
- * plot-area / legend / data-table `<c:spPr>` border slots — same hex
- * grammar, same `<a:ln>` slot on the `CT_ShapeProperties` schema —
- * but lands on `<c:dLbls>`'s own `<c:spPr>` block.
- */
-function resolveDataLabelsBorderColor(value: string | undefined): string | undefined {
-  return normalizeTitleColor(value);
-}
-
-/**
- * Build the optional `<c:spPr>` block inside `<c:dLbls>`. Surfaces the
- * solid fill color knob ({@link ChartDataLabels.fillColor}) and the
- * border (line) color knob ({@link ChartDataLabels.borderColor}) —
- * every other `<c:spPr>` child (`<a:effectLst>` effects, gradient /
- * pattern / picture fills, line dash / width / compound styles) is
- * intentionally not modelled at this layer.
- *
- * Returns `undefined` when both fields are unset / malformed so the
- * writer skips the entire `<c:spPr>` block — an empty `<c:spPr/>`
- * collapses to the inherited theme fill / stroke Excel picks anyway,
- * and omitting it keeps untouched chart XML byte-clean. When at least
- * one knob lands on the wire, the children are emitted in
- * `CT_ShapeProperties` schema order: `<a:solidFill>` (fill) then
- * `<a:ln>` (line / stroke).
- *
- * The emitted block mirrors the minimal `<c:spPr>` shape Excel writes
- * when the user pins "Format Data Labels -> Fill -> Solid fill ->
- * Color" and / or "Format Data Labels -> Border -> Solid line ->
- * Color": `<c:spPr><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill><a:ln><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill></a:ln></c:spPr>`. The `val` attribute holds the
- * canonical 6-character uppercase hex form (the writer normalizes the
- * input ahead of this call so a malformed source value never reaches
- * emit).
- *
- * Mirrors {@link buildLegendSpPr} / {@link buildDataTableSpPr} but
- * on a distinct host element — the data-labels fill / stroke paint
- * the background and outline of each label box, while the legend /
- * data-table variants paint a different chart-frame element.
- */
-function buildDataLabelsSpPr(
-  fillRgbHex: string | undefined,
-  borderRgbHex: string | undefined,
-  borderWidthPt: number | undefined,
-  borderDash: ChartBorderDash | undefined,
-): string | undefined {
-  if (
-    fillRgbHex === undefined &&
-    borderRgbHex === undefined &&
-    borderWidthPt === undefined &&
-    borderDash === undefined
-  ) {
-    return undefined;
-  }
-  const children: string[] = [];
-  if (fillRgbHex !== undefined) {
-    children.push(
-      xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: fillRgbHex })]),
-    );
-  }
-  if (borderRgbHex !== undefined || borderWidthPt !== undefined || borderDash !== undefined) {
-    const lnAttrs: Record<string, string | number> = {};
-    if (borderWidthPt !== undefined) {
-      // OOXML stores stroke width in EMU (1 pt = 12 700 EMU). Round to
-      // the nearest integer because the schema types `w` as `xsd:int`.
-      lnAttrs.w = Math.round(borderWidthPt * EMU_PER_PT);
-    }
-    const lnChildren: string[] = [];
-    if (borderRgbHex !== undefined) {
-      lnChildren.push(
-        xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: borderRgbHex })]),
-      );
-    }
-    // `<a:prstDash>` follows `<a:solidFill>` per CT_LineProperties
-    // schema sequence (ECMA-376 Part 1, §20.1.2.3.24).
-    if (borderDash !== undefined) {
-      lnChildren.push(xmlSelfClose("a:prstDash", { val: borderDash }));
-    }
-    children.push(
-      lnChildren.length === 0
-        ? xmlSelfClose("a:ln", lnAttrs)
-        : xmlElement("a:ln", Object.keys(lnAttrs).length > 0 ? lnAttrs : undefined, lnChildren),
-    );
-  }
-  return xmlElement("c:spPr", undefined, children);
-}
-
-/**
- * Build the `<c:txPr>` block that carries a data-label's typography
- * pins. Returns `undefined` when every input is unset so the caller
- * can elide the element entirely (Excel's reference serialization
- * omits `<c:txPr>` from `<c:dLbls>` when the labels render at the
- * theme-default style).
- *
- * The emitted block mirrors the minimal `<c:txPr>` shape Excel writes
- * when the user pins a data-label typography knob — `<a:bodyPr/>` (no
- * rotation because the data-label rotation is parked in a separate
- * extension element Excel emits at write time), `<a:lstStyle/>` is
- * the empty list-style placeholder the schema requires, and the
- * `<a:p><a:pPr><a:defRPr ...><a:solidFill>...</a:solidFill></a:defRPr>
- * </a:pPr><a:endParaRPr/></a:p>` paragraph stub Excel always emits
- * hosts the typography attributes on `<a:defRPr>`. The `<a:defRPr>`
- * element expands from self-closing to wrapping a single
- * `<a:solidFill>` child when a color is set; otherwise the writer
- * keeps the existing self-closing form so a fresh chart with no
- * custom color matches Excel's reference serialization byte-for-byte.
- * The bold and italic flags each emit a literal `b="1"` / `b="0"`
- * (or `i="1"` / `i="0"`) whenever the input is a boolean — `false`
- * pins the OOXML default explicitly, which is functionally identical
- * to absence but lets a clone target override an upstream `b="1"` /
- * `i="1"` from a templated chart. The underline flag emits `u="sng"`
- * (single underline — Excel's UI variant) for `true` and `u="none"`
- * (the OOXML default) for `false`, with the same override semantics
- * as bold and italic. The strikethrough flag rides as
- * `strike="sngStrike"` on the same `<a:defRPr>` slot when the input
- * is `true`; absence (and explicit `false`, which the resolver
- * collapses to `undefined`) skips the attribute entirely since the
- * OOXML default `"noStrike"` is functionally identical to absence.
- * Mirrors the chart-title / axis-title / axis tick-label / legend
- * `<c:txPr>` slots exactly so a re-parse picks the value off the
- * canonical default-paragraph slot every other typography reader
- * expects.
- */
-function buildDataLabelsTxPr(
-  fontSizePt: number | undefined,
-  rgbHex: string | undefined,
-  bold: boolean | undefined,
-  italic: boolean | undefined,
-  underline: boolean | undefined,
-  strikethrough: boolean | undefined,
-  fontFamily: string | undefined,
-): string | undefined {
-  if (
-    fontSizePt === undefined &&
-    rgbHex === undefined &&
-    bold === undefined &&
-    italic === undefined &&
-    underline === undefined &&
-    strikethrough === undefined &&
-    fontFamily === undefined
-  )
-    return undefined;
-  const defRPrAttrs: Record<string, string | number> = {};
-  if (fontSizePt !== undefined) defRPrAttrs.sz = fontSizePt * TITLE_FONT_SZ_PER_POINT;
-  if (bold !== undefined) defRPrAttrs.b = bold ? 1 : 0;
-  if (italic !== undefined) defRPrAttrs.i = italic ? 1 : 0;
-  if (underline !== undefined) defRPrAttrs.u = underline ? "sng" : "none";
-  // Strikethrough rides as `strike="sngStrike"` on the same
-  // `<a:defRPr>` slot. Absence collapses to omitting the attribute
-  // entirely (the OOXML default `"noStrike"` is functionally
-  // identical to absence — the reader collapses both to `undefined`).
-  // The writer never emits `"noStrike"` or `"dblStrike"` so the
-  // surfaced shape stays consistent with Excel's UI checkbox.
-  if (strikethrough === true) defRPrAttrs.strike = "sngStrike";
-  // OOXML's `<a:defRPr><a:solidFill><a:srgbClr val="RRGGBB"/>
-  // </a:solidFill></a:defRPr>` carries the data-label font color.
-  // Absence (`undefined`) collapses to skipping the `<a:solidFill>`
-  // child entirely so the labels inherit the theme text color
-  // (Excel's reference behavior for fresh data labels that have not
-  // had a custom color picked).
-  const solidFillChild = rgbHex
-    ? xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: rgbHex })])
-    : undefined;
-  // OOXML's `<a:defRPr><a:latin typeface=".."/></a:defRPr>` carries
-  // the data-label font family. The `<a:latin>` element follows
-  // `<a:solidFill>` per the CT_TextCharacterProperties child sequence
-  // (ECMA-376 Part 1, §21.1.2.3.7). Absence (`undefined`) collapses
-  // to omitting the entire `<a:latin>` element so the labels inherit
-  // the theme typeface (Excel's reference behavior for fresh data
-  // labels that have not had a custom font picked).
-  const latinChild = fontFamily ? xmlSelfClose("a:latin", { typeface: fontFamily }) : undefined;
-  // When a fill color or a typeface is set the `<a:defRPr>` slot
-  // expands from self-closing to wrapping the children; otherwise the
-  // writer keeps the existing self-closing form so a fresh chart with
-  // no custom color or font matches Excel's reference serialization
-  // byte-for-byte. Children are emitted in
-  // CT_TextCharacterProperties' canonical schema order: solidFill
-  // first, then latin.
-  const defRPrChildren: string[] = [];
-  if (solidFillChild) defRPrChildren.push(solidFillChild);
-  if (latinChild) defRPrChildren.push(latinChild);
-  const defRPr =
-    defRPrChildren.length > 0
-      ? xmlElement("a:defRPr", defRPrAttrs, defRPrChildren)
-      : xmlSelfClose("a:defRPr", defRPrAttrs);
-  return xmlElement("c:txPr", undefined, [
-    xmlSelfClose("a:bodyPr"),
-    xmlSelfClose("a:lstStyle"),
-    xmlElement("a:p", undefined, [
-      xmlElement("a:pPr", undefined, [defRPr]),
-      xmlSelfClose("a:endParaRPr", { lang: "en-US" }),
-    ]),
-  ]);
-}
 
 // ── Manual Layout ────────────────────────────────────────────────────
